reward model extraction base class
