+++ test/registered/ascend/llm_models/test_npu_ernie_4_5_21b_a3b_pt.py
@@ -9,7 +9,6 @@ register_npu_ci(
     est_time=400,
     suite="full-1-npu-a3",
     nightly=True,
-    disabled="https://github.com/Ascend/sglang/issues/35",
 )
