+++ test/registered/ascend/llm_models/test_npu_gpt_oss_120b_bf16.py
@@ -9,7 +9,6 @@ register_npu_ci(
     est_time=400,
     suite="nightly-8-npu-a3",
     nightly=True,
-    disabled="https://github.com/Ascend/sglang/issues/122",
 )
