+++ test/registered/ascend/llm_models/test_npu_grok_2.py
@@ -12,7 +12,6 @@ register_npu_ci(
     est_time=400,
     suite="full-16-npu-a3",
     nightly=False,
-    disabled="https://github.com/Ascend/sglang/issues/24",
 )
