+++ test/registered/ascend/llm_models/test_npu_trinity_mini.py
@@ -9,7 +9,6 @@ register_npu_ci(
     est_time=400,
     suite="full-2-npu-a3",
     nightly=True,
-    disabled="https://github.com/Ascend/sglang/issues/125",
 )
