+++ test/registered/ascend/vlm_models/test_npu_dots_ocr.py
@@ -8,7 +8,6 @@ register_npu_ci(
     est_time=400,
     suite="nightly-1-npu-a3",
     nightly=True,
-    disabled="https://github.com/Ascend/sglang/issues/81",
 )
