+++ test/registered/ascend/vlm_models/test_npu_llava_v1_6_34b.py
@@ -11,7 +11,6 @@ register_npu_ci(
     est_time=400,
     suite="nightly-4-npu-a3",
     nightly=True,
-    disabled="run failed",
 )
